fix: removed .isRequired from genre.name and director.name in MovieCard PropTypes
Removed .isRequired constraint due to a PropTypes warning in the MovieCard and MovieView components. The warning was caused by movie.genre[0].name and movie.director[0].name being undefined, even though they were marked as required in PropTypes. This change was made because the genre and director arrays are already required, making it unnecessary to also require their name properties.